[huny.jeong] 로또 수동 - 2단계 (#57)
* feat : 로또 수동입력 추가

* refactor : 로또 자동 생성 부분 factory 클래스로 분리

* feat : factory 클래스 로또 수동 생성 추가

* refactor : Lottos getLottos 불변으로 반환하기

* feat : 수동 생성 로또와 자동 생성 로또 합치기

* feat : 수동 생성 로또개수 예외처리

* feat : 컨트롤러에 로또 수동 생성 추가

* feat : WinningNumbers 클래스로 지난주 당첨 번호, 보너스볼 묶기

* feat : RankingResult 일급 콜렉션 생성

* refactor : RankingResult 일급 콜렉션으로 묶기

* refactor : OutputView 줄바꿈 처리

* refactor : Reformat code

* refactor : 로또 수동, 자동 생성 팩토리 분리

* refactor : 테스트 코드 수정

* refactor : 수동/자동 로또 번호 합치기 수정

+++ src/main/java/lotto/model/enums/RankingType.java
@@ -10,7 +10,7 @@ public enum RankingType {
     FIFTH(5000, 3, false, "3개 일치 (5000원)"),
     DEFAULT(0, 0, false, "");
 
-    private final int prize;
+    private final long prize;
     private final int count;
     private final boolean hasBonusNumber;
 
@@ -30,11 +30,11 @@ public enum RankingType {
                 .orElse(DEFAULT);
     }
 
-    public int getPrize() {
+    public long getPrize() {
         return prize;
     }
 
-    public String getHowManyMatches(){
+    public String getHowManyMatches() {
         return howManyMatches;
     }
 }
